--- a/ProtoPromise/nuget/ProtoPromise.nuspec
+++ b/ProtoPromise/nuget/ProtoPromise.nuspec
@@ -21,7 +21,7 @@
       <group targetFramework="net40" />
       <group targetFramework="net45" />
       <group targetFramework="net47" />
-      <!--<group targetFramework="net5.0" />-->
+      <group targetFramework="net5.0" />
     </dependencies>
 
     <readme>readme.md</readme>
@@ -29,18 +29,25 @@
 
   <files>
     <file src="targets\**" target="build" />
+    
     <file src="..\bin\Release\netstandard2.0\**" target="lib/netstandard2.0/Release" />
     <file src="..\bin\Debug\netstandard2.0\**" target="lib/netstandard2.0/Debug" />
+    
     <file src="..\bin\Release\net35\**" target="lib/net35/Release" />
     <file src="..\bin\Debug\net35\**" target="lib/net35/Debug" />
+    
     <file src="..\bin\Release\net40\**" target="lib/net40/Release" />
     <file src="..\bin\Debug\net40\**" target="lib/net40/Debug" />
+    
     <file src="..\bin\Release\net45\**" target="lib/net45/Release" />
     <file src="..\bin\Debug\net45\**" target="lib/net45/Debug" />
+    
     <file src="..\bin\Release\net47\**" target="lib/net47/Release" />
     <file src="..\bin\Debug\net47\**" target="lib/net47/Debug" />
-    <!--<file src="..\bin\Release\net5.0\**" target="lib/net5.0/Release" />-->
-    <!--<file src="..\bin\Debug\net5.0\**" target="lib/net5.0/Debug" />-->
+    
+    <file src="..\bin\Release\net5.0\**" target="lib/net5.0/Release" />
+    <file src="..\bin\Debug\net5.0\**" target="lib/net5.0/Debug" />
+    
     <file src="readme.md" target="" />
   </files>
 </package>