--- a/ProtoPromise/nuget/targets/net5.0/ProtoPromise.targets
+++ b/ProtoPromise/nuget/targets/net5.0/ProtoPromise.targets
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Choose>
+    <When Condition="$(Configuration.Contains('Release'))">
+      <ItemGroup>
+        <Reference Include="ProtoPromise">
+          <HintPath>$(MSBuildThisFileDirectory)..\..\lib\net5.0\Release\ProtoPromise.dll</HintPath>
+        </Reference>
+      </ItemGroup>
+    </When>
+    <Otherwise>
+      <ItemGroup>
+        <Reference Include="ProtoPromise">
+          <HintPath>$(MSBuildThisFileDirectory)..\..\lib\net5.0\Debug\ProtoPromise.dll</HintPath>
+        </Reference>
+      </ItemGroup>
+    </Otherwise>
+  </Choose>
+</Project>