--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
@@ -22,7 +22,7 @@ namespace Proto.Promises
 
         internal interface IRejectionToContainer
         {
-            ValueContainer ToContainer(ITraceable traceable);
+            RejectContainer ToContainer(ITraceable traceable);
         }
 
         internal interface ICantHandleException