--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
@@ -15,11 +15,6 @@ namespace Proto.Promises
             T Next { get; set; }
         }
 
-        internal interface IProgressInvokable : ILinked<IProgressInvokable>
-        {
-            void Invoke(ref ExecutionScheduler executionScheduler);
-        }
-
         internal interface IRejectionToContainer
         {
             RejectContainer ToContainer(ITraceable traceable);