--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/PoolInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/PoolInternal.cs
@@ -58,7 +58,7 @@ namespace Proto.Promises
 
                 private static void Clear()
                 {
-                    s_pool.Clear();
+                    s_pool.ClearUnsafe();
                 }
 
                 [MethodImpl(InlineOption)]