Lots of small changes to squeeze performance (check each commit message for details).

Closes #26.

Master:

|         Type | Pending |       Mean |    Error |    StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|----------:|----------:|---------:|
|   AsyncAwait |   False |   915.5 ns | 16.59 ns |  14.71 ns |         - |        - |
| ContinueWith |   False |   954.1 ns | 17.83 ns |  17.51 ns |         - |        - |
|   AsyncAwait |    True | 4,270.5 ns | 83.69 ns | 105.84 ns |         - |    864 B |
| ContinueWith |    True | 4,297.5 ns | 39.35 ns |  34.88 ns |         - |    560 B |

This PR:

|         Type | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |   False |   612.8 ns | 11.28 ns | 10.55 ns |         - |        - |
| ContinueWith |   False |   904.7 ns | 11.95 ns |  9.98 ns |         - |        - |
|   AsyncAwait |    True | 2,830.3 ns | 37.02 ns | 32.82 ns |         - |    712 B |
| ContinueWith |    True | 3,081.5 ns | 22.71 ns | 18.97 ns |         - |    416 B |